// src/modules/address/enums/address-callback-data.enum.ts
export enum AddressMenuCallbackData {
  CreateNew = 'create_new_address',
  List = 'list_of_addresses',
  Remove = 'remove_address_',
  Update = 'update_address_',
  SetDefault = 'set_default_address_',
  BackToMain = 'back_to_main_address',
  SkipUnit = 'skip_unit_address',
  UseProfileFullName = 'use_profile_full_name',
  ShowDetails = 'show_address_details_',
  CloseMainMenu = 'close_main_menu',
  UpdateFullName = 'update_address_full_name',
  UpdateProvince = 'update_address_province',
  UpdateCity = 'update_address_city',
  UpdateBuildingNumber = 'update_address_building_number',
  UpdatePostalCode = 'update_address_postal_code',
  UpdatePostalAddress = 'update_address_postal_address',
  UpdateUnit = 'update_address_unit',
  CancelUpdate = 'cancel_update_address',
  UnsetDefault = 'unset_default_address_',
}